requirements.txt
numpy
yaml
pynput
vosk
pyAudioAnalysis
python-dotenv
twitchio
aiohttp
pyaudio